Extract local content entity enum to separate file
