+++ WordPress/src/main/java/org/wordpress/android/localcontentmigration/LocalMigrationContentProvider.kt
@@ -6,7 +6,6 @@ import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
 import dagger.hilt.android.EntryPointAccessors
 import dagger.hilt.components.SingletonComponent
-import org.wordpress.android.fluxc.model.PostModel
 import org.wordpress.android.localcontentmigration.LocalContentEntity.AccessToken
 import org.wordpress.android.localcontentmigration.LocalContentEntity.EligibilityStatus
 import org.wordpress.android.localcontentmigration.LocalContentEntity.Post
@@ -60,34 +59,4 @@ class LocalMigrationContentProvider: TrustedQueryContentProvider() {
 
 interface LocalDataProviderHelper {
     fun getData(localSiteId: Int? = null, localEntityId: Int? = null): LocalContentEntityData
-}
-
-enum class LocalContentEntity(private val isSiteContent: Boolean = false) {
-    EligibilityStatus,
-    AccessToken,
-    Site,
-    Post(isSiteContent = true),
-    ;
-
-    open val contentIdCapturePattern = when (isSiteContent) {
-        true -> Regex("site/(\\d+)/${name}(?:/(\\d+))?")
-        false -> Regex(name)
-    }
-
-    open fun getPathForContent(localSiteId: Int?, localEntityId: Int?) = when (this.isSiteContent) {
-        true -> "site/${localSiteId}/${name}${ localEntityId?.let { "/${it}" } ?: "" }"
-        false -> name
-    }
-}
-
-sealed class LocalContentEntityData {
-    data class EligibilityStatusData(
-        val isEligible: Boolean,
-        val siteCount: Int,
-    ): LocalContentEntityData()
-
-    data class AccessTokenData(val token: String): LocalContentEntityData()
-    data class SitesData(val localIds: List<Int>): LocalContentEntityData()
-    data class PostsData(val localIds: List<Int>): LocalContentEntityData()
-    data class PostData(val post: PostModel) : LocalContentEntityData()
 }
